# dimos/server/openapi.json
{
  "components": {
    "schemas": {
...
      "BlueprintList": {
        "additionalProperties": true,
        "properties": {
          "blueprints": {
            "description": "Built-in blueprints (sorted, without demo-*), then external ones",
            "items": {
              "$ref": "#/components/schemas/BlueprintName"
            },
            "title": "Blueprints",
            "type": "array"
          }
        },
        "required": [
          "blueprints"
        ],
        "title": "BlueprintList",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": true,
        "description": "Every error answer (4xx and 5xx).",
        "properties": {
          "error": {
            "description": "What went wrong, in words a person can act on",
            "examples": [
              "no upload u7"
            ],
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The dimos server's HTTP API: blueprints, global config, runs and their logs, events, Dimensional cloud uploads,\ndiscovery (blueprints, modules, message types), docs, extras and jobs.\n\ndimOS Desktop starts the server (`python -m dimos.server`, on a unix socket) and forwards `/dimos/...` to it unchanged.\nEvery answer is JSON unless the operation says otherwise; every error is `{\"error\": \"<message>\"}` (ErrorResponse).\nState is HTTP; changes are events, published on zenoh at `<ns>/dimos/events/<type>` (the `events` tag).\n\nExtensions: `x-family: dimos` on every operation; `x-agent: true` marks what Desktop's agent finds through\nsearch_endpoints; `x-mcp-tool` names an MCP tool that does the same; `x-zenoh-key` on an event schema is its key.\n",
    "title": "dimos server",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/dimos/blueprints": {
      "get": {
        "description": "What `dimos list` prints: built-in blueprints (without demo-*), then external ones from installed packages. Cached for 60 s; `fresh` refills the cache first. No other side effects. `importable`, `import_error` and `missing_module` come from the discovery cache (null until the scan reaches the blueprint: GET /dimos/discovery).",
        "operationId": "get_dimos_blueprints",
        "parameters": [
          {
            "description": "check again now instead of answering from the cache",
            "in": "query",
            "name": "fresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "check again now instead of answering from the cache",
              "title": "Fresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintList"
                }
              }
            },
            "description": "`{ blueprints: [{ name, kind: \"builtin\"|\"external\", importable, import_error, missing_module }] }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the server refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Every blueprint dimos can run (name, builtin/external) and whether it imports",
        "tags": [
          "blueprints"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/catalog": {
      "get": {
        "description": "Imports every built-in blueprint and module in a child process (180 s timeout; cached 10 min) and lists blueprints with their robot and modules, modules with their streams and skills, and every skill with its parameters. What fails to import is listed in `errors`; the rest still answers.",
        "operationId": "get_dimos_catalog",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            },
            "description": "`{ blueprints: [{ name, ref, robot, modules }], modules: [{ name, class, doc, robots, inputs, outputs, skills }], skills: [{ name, doc, params, module, robots }], errors }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Every blueprint, module and skill (imports them all, in a child process: slow the first time)",
        "tags": [
          "blueprints"
        ],
        "x-agent": false,
        "x-family": "dimos"
      }
    },
    "/dimos/cloud/account": {
      "get": {
        "description": "Asks Dimensional cloud who the stored key (or DIMOS_API_KEY) belongs to, in a child process (90 s timeout). Cached for 20 s; `fresh` asks again. A logged-in answer restarts an upload queue that was waiting for a login.",
        "operationId": "get_dimos_cloud_account",
        "parameters": [
          {
            "description": "check again now instead of answering from the cache",
            "in": "query",
            "name": "fresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "check again now instead of answering from the cache",
              "title": "Fresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "`{ loggedIn, email, scopes, source, cloudUrl, error }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the server refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Whether this machine is logged in to Dimensional cloud, and as whom",
        "tags": [
          "cloud"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
    "/dimos/cloud/login": {
      "delete": {
        "description": "Kills the login's child process; a starting or pending login goes back to idle. Answers the login state.",
        "operationId": "delete_dimos_cloud_login",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Login"
                }
              }
            },
            "description": "`Login`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Cancel the pending cloud login",
        "tags": [
          "cloud"
        ],
        "x-agent": false,
        "x-family": "dimos"
      },
      "get": {
        "description": "The device login's state; a pending one past its expiry turns expired. No other side effects.",
        "operationId": "get_dimos_cloud_login",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Login"
                }
              }
            },
            "description": "`Login`: `{ state, url, urlComplete, code, expiresAt, email, error }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "The cloud login in progress: state (idle, starting, pending, approved, denied, expired, failed), url, code",
        "tags": [
          "cloud"
        ],
        "x-agent": true,
        "x-family": "dimos"
      },
      "post": {
        "description": "Starts dimos's device login in a child process (or returns the one already waiting) and answers once the code is known (pending), within 30 s. Show the URL and code; `cloud-login` events (or GET /dimos/cloud/login) follow it. Once approved, dimos stores the key and a waiting upload queue goes on.",
        "operationId": "post_dimos_cloud_login",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Login"
                }
              }
            },
            "description": "`Login`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Start logging this machine in to Dimensional cloud: returns a URL and a code the user approves in any signed-in browser",
        "tags": [
          "cloud"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/cloud/logout": {
      "post": {
        "description": "Forgets the stored cloud key (dimos's `logout`) and resets the login, then answers the account as GET /dimos/cloud/account?fresh=1 would.",
        "operationId": "post_dimos_cloud_logout",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "`{ loggedIn, email, scopes, source, cloudUrl, error }`, logged out"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Log this machine out of Dimensional cloud",
        "tags": [
          "cloud"
        ],
        "x-agent": false,
        "x-family": "dimos"
      }
    },
...
    "/dimos/global-config": {
      "get": {
        "description": "GlobalConfig's JSON Schema and defaults (cached 10 min) and Desktop's saved overrides from config.yaml `dimos.global_config`. No side effects.",
        "operationId": "get_dimos_global_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalConfig"
                }
              }
            },
            "description": "`{ schema, defaults, overrides }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "dimos GlobalConfig: JSON schema, defaults, Desktop's overrides",
        "tags": [
          "global-config"
        ],
        "x-agent": true,
        "x-family": "dimos"
      },
      "put": {
        "description": "Replaces config.yaml's `dimos.global_config` with `overrides` (null values dropped), keeping the rest of the file. Takes effect at the next launch; a running blueprint is untouched. 400 for a key that isn't a GlobalConfig field `dimos` takes as a flag, or a value GlobalConfig refuses. Answers like GET.",
        "operationId": "put_dimos_global_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalConfig"
                }
              }
            },
            "description": "`{ schema, defaults, overrides }`, with the saved overrides"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the server refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Save Desktop's GlobalConfig overrides (null removes one); they become `--key=value` on every launch",
        "tags": [
          "global-config"
        ],
        "x-agent": false,
        "x-family": "dimos"
      }
    },
...
    "/dimos/info": {
      "get": {
        "description": "Reads the checkout's pyproject.toml and whether its `.venv/bin/dimos` exists, and checks the version against the range Desktop supports ($DESKTOP_DIMOS_RANGE). A launch is refused while `inRange` is false. No side effects.",
        "operationId": "get_dimos_info",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            },
            "description": "`{ dir, found, installed, version, range, inRange }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "The dimos checkout this server drives: dir, version, installed",
        "tags": [
          "server"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/paths": {
      "get": {
        "description": "Absolute paths, read from the environment and Desktop's config.yaml. Desktop compares `dimosDir` with its configured checkout to tell whether this server is the right one, and `server` with its own binary to tell whether its built-in server is outdated (this server runs python, so it never is). No side effects.",
        "operationId": "get_dimos_paths",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paths"
                }
              }
            },
            "description": "`{ dimosDir, runsDir, logsDirs, recordingsDir, server: { exe, exeModified } }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Where dimos keeps its things: the checkout, run registry, log folders, recordings folder",
        "tags": [
          "server"
        ],
        "x-agent": false,
        "x-family": "dimos"
      }
    },
...
  },
...
}